New features newSim and loadSim
Revamped the backend and frontend to include new Sim and loadSim which utilized a unique id and name to either create or load a simulation. This id is present in 3 tables and helps seperate data and makes it possible to alter the data freely in a different way in new simulations while preserving the original data for future simulations. I also added a delete function to delete entries but have not implemented yet in the frontend.